refactor: extract data queries and navigation into custom hooks
Move all data queries, derived data computation, and navigation state
out of the Timeline component into two focused hooks:

- useTimelineNavigation: signals, view state, zoom handling, date range
  expansion, jump/reset, bucket sizes, view label (~130 lines)
- useTimelineData: 12 useQuery calls, 15+ useMemo derivations, chart
  item filtering and lane packing (~280 lines)

index.tsx: 1738 → 1320 lines (-418)
