+++ apps/web/src/pages/Timeline/drawVerticalHelpers.ts
@@ -1,4 +1,5 @@
 import type * as d3 from 'd3'
+
 import { formatISO } from 'date-fns'
 
 import type { ChartItem, Column } from './types'
