+++ apps/web/src/pages/Timeline/useTimelineNavigation.ts
@@ -1,0 +1,147 @@
+import { signal } from '@preact/signals'
+import { addDays, differenceInCalendarDays, endOfDay, format, formatISO, startOfDay, subDays } from 'date-fns'
+import { useCallback, useMemo } from 'preact/hooks'
+
+import { getDefaultViewEnd, getDefaultViewStart, parseViewHash } from './viewHash'
+
+// ── Signals (module-level, persist across SPA navigations) ────────────────────
+
+const fromDate = signal(formatISO(subDays(new Date(), 1), { representation: 'date' }))
+const toDate = signal(formatISO(new Date(), { representation: 'date' }))
+const viewStart = signal<Date | null>(null)
+const viewEnd = signal<Date | null>(null)
+
+// Initialise signals from hash on page load
+const _initialHash = parseViewHash()
+if (_initialHash.from) {
+  viewStart.value = _initialHash.from
+  viewEnd.value = _initialHash.to
+  const fetchFrom = _initialHash.from
+  const fetchTo = _initialHash.to ?? _initialHash.from
+  fromDate.value = formatISO(subDays(fetchFrom, 1), { representation: 'date' })
+  const todayStr = formatISO(new Date(), { representation: 'date' })
+  const expandedTo = formatISO(addDays(fetchTo, 1), { representation: 'date' })
+  toDate.value = expandedTo > todayStr ? todayStr : expandedTo
+}
+
+export { _initialHash }
+
+export interface TimelineNavigation {
+  effectiveViewStart: Date
+  effectiveViewEnd: Date
+  fetchStart: Date
+  fetchEnd: Date
+  fromDate: typeof fromDate
+  toDate: typeof toDate
+  viewStart: typeof viewStart
+  viewEnd: typeof viewEnd
+  handleZoom: (start: Date, end: Date) => void
+  handleJumpDays: (days: number) => void
+  handleResetToToday: () => void
+  viewLabel: string
+  bucketSize: string
+  barBucketSize: '1h' | '1d' | '1w'
+  screentimeMergeGapMs: number
+}
+
+export const useTimelineNavigation = (): TimelineNavigation => {
+  const effectiveViewStart = viewStart.value ?? getDefaultViewStart()
+  const effectiveViewEnd = viewEnd.value ?? getDefaultViewEnd()
+
+  const fetchStart = startOfDay(new Date(fromDate.value))
+  const fetchEnd = endOfDay(new Date(toDate.value))
+
+  const screentimeMergeGapMs = useMemo(() => {
+    const days = differenceInCalendarDays(new Date(toDate.value), new Date(fromDate.value))
+    if (days > 50) return 4 * 60 * 60 * 1000
+    if (days > 2) return 60 * 60 * 1000
+    return 10 * 60 * 1000
+  }, [fromDate.value, toDate.value])
+
+  const bucketSize = useMemo(() => {
+    const days = differenceInCalendarDays(effectiveViewEnd, effectiveViewStart)
+    if (days > 500) return '1w'
+    if (days > 21) return '1d'
+    if (days > 5) return '1h'
+    if (days > 1) return '15m'
+    return '5m'
+  }, [effectiveViewStart, effectiveViewEnd])
+
+  const barBucketSize = useMemo((): '1h' | '1d' | '1w' => {
+    const days = differenceInCalendarDays(effectiveViewEnd, effectiveViewStart)
+    if (days > 50) return '1w'
+    if (days > 2) return '1d'
+    return '1h'
+  }, [effectiveViewStart, effectiveViewEnd])
+
+  const handleZoom = useCallback((zoomStart: Date, zoomEnd: Date) => {
+    viewStart.value = zoomStart
+    viewEnd.value = zoomEnd
+
+    const currentFetchStart = startOfDay(new Date(fromDate.value))
+    const currentFetchEnd = endOfDay(new Date(toDate.value))
+    const todayStr = formatISO(new Date(), { representation: 'date' })
+
+    let needsExpand = false
+    let newFrom = fromDate.value
+    let newTo = toDate.value
+
+    if (zoomStart < currentFetchStart) {
+      newFrom = formatISO(subDays(zoomStart, 3), { representation: 'date' })
+      needsExpand = true
+    }
+    if (zoomEnd > currentFetchEnd) {
+      const expanded = formatISO(addDays(zoomEnd, 3), { representation: 'date' })
+      newTo = expanded > todayStr ? todayStr : expanded
+      needsExpand = true
+    }
+
+    if (needsExpand) {
+      fromDate.value = newFrom
+      toDate.value = newTo
+    }
+  }, [])
+
+  const handleJumpDays = useCallback(
+    (days: number) => {
+      const currentStart = viewStart.value ?? getDefaultViewStart()
+      const currentEnd = viewEnd.value ?? getDefaultViewEnd()
+      const newStart = addDays(currentStart, days)
+      const newEnd = addDays(currentEnd, days)
+      const todayEnd = endOfDay(new Date())
+      if (newEnd > todayEnd) return
+      handleZoom(newStart, newEnd)
+    },
+    [handleZoom],
+  )
+
+  const handleResetToToday = useCallback(() => {
+    viewStart.value = null
+    viewEnd.value = null
+    fromDate.value = formatISO(subDays(new Date(), 1), { representation: 'date' })
+    toDate.value = formatISO(new Date(), { representation: 'date' })
+  }, [])
+
+  const viewLabel =
+    format(effectiveViewStart, 'MMM d') === format(effectiveViewEnd, 'MMM d')
+      ? format(effectiveViewStart, 'MMM d, yyyy')
+      : `${format(effectiveViewStart, 'MMM d')} – ${format(effectiveViewEnd, 'MMM d, yyyy')}`
+
+  return {
+    barBucketSize,
+    bucketSize,
+    effectiveViewEnd,
+    effectiveViewStart,
+    fetchEnd,
+    fetchStart,
+    fromDate,
+    handleJumpDays,
+    handleResetToToday,
+    handleZoom,
+    screentimeMergeGapMs,
+    toDate,
+    viewEnd,
+    viewLabel,
+    viewStart,
+  }
+}
